add ssize_t to return size in functions that can fail

+++ src/lib/bitmap.c
@@ -15,15 +15,15 @@
 
 #include <bitmap.h>
 
-long int bitmap_find_nth(bitmap_t map, size_t size, size_t nth, size_t start,
+ssize_t bitmap_find_nth(bitmap_t map, size_t size, size_t nth, size_t start,
                         bool set)
 {
     if (size <= 0 || nth <= 0 || start < 0) return -1;
 
     size_t count = 0;
-    uint64_t bit = set ? 1 : 0;
+    unsigned int bit = set ? 1 : 0;
 
-    for (long int i = start; i < size; i++) {
+    for (ssize_t i = start; i < size; i++) {
         if (bitmap_get(map, i) == bit) {
             if (++count == nth) return i;
         }
@@ -37,7 +37,7 @@ size_t bitmap_count_consecutive(bitmap_t map, size_t size, size_t start,
 {
     if (n <= 1) return n;
 
-    uint8_t b = bitmap_get(map, start);
+    unsigned int b = bitmap_get(map, start);
     size_t count = 1;
     start += 1;
 
@@ -53,11 +53,11 @@ size_t bitmap_count_consecutive(bitmap_t map, size_t size, size_t start,
     return count;
 }
 
-int bitmap_find_consec(bitmap_t map, size_t size, size_t start, size_t n,
+ssize_t bitmap_find_consec(bitmap_t map, size_t size, size_t start, size_t n,
                             bool set)
 {
-    int count = 0;
-    int i = 0;
+    ssize_t count = 0;
+    ssize_t i = 0;
 
     // find first set
     if ((i = bitmap_find_nth(map, size, 1, start, set)) < 0) return -1;
